services/dave/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.24.0
pydantic==2.4.0
python-json-logger==2.0.7
opentelemetry-sdk==1.26.0
opentelemetry-api==1.26.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
# Dave Specific
sse-starlette>=1.8.0
google-generativeai>=0.5.0
supabase>=2.4.0
slowapi>=0.1.9
beautifulsoup4>=4.12.0
tenacity>=8.2.0
